Un-ban C++ <regex> header
We need to use _a_ regular expression library, it'll either be <regex>
from the standard library, or RE2. The problem with <regex> is that
cpplint bans it (fine!), but the problem with RE2 is that it'll
introduce its own tests into our test suite, making it a hassle to
vendor.

We were forced to make a difficult choice:
  0. Modify a vendored file (cpplint.py), pray that it'll still be easy
    to merge upstream updates; or
  0. include RE2's tests in our test suite.

We are going with the C++11 standard library in this case. Until RE2
makes the vendoring story more palatable.